# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Titre de votre API",
        "description": "Description de votre API",
        "contact": {
            "email": "support@votreapi.com"
        },
        "license": {
            "name": "Licence de votre API",
            "url": "URL de la licence"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Liste des clients",
                "description": "Récupère la liste des clients, avec possibilité de filtrer par la présence d'un compte utilisateur et par l'état d'activation du compte.",
                "operationId": "b302d909c0e1166a2c033bfd17ed030f",
                "parameters": [
                    {
                        "name": "comptes",
                        "in": "query",
                        "description": "Filtrer par présence d'un compte utilisateur (oui/non)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrer par état d'activation du compte (oui/non)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des clients récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Créer un client",
                "description": "Crée un nouveau client avec la possibilité d'associer un compte utilisateur.",
                "operationId": "72a51329a9c85d73e15e640b5b53e470",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la création du client"
                    }
                }
            }
        },
        "/api/v1/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Afficher un client",
                "description": "Récupère les informations d'un client par son ID.",
                "operationId": "0a628c8462ccec3d2798255b1015c22c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations du client récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                }
            }
        },
        "/api/v1/clients/telephone/{telephone}": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Afficher un client par téléphone",
                "description": "Récupère les informations d'un client en fonction de son numéro de téléphone.",
                "operationId": "27e73f00813ba2b42f09781d21a5b064",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informations du client récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                }
            }
        },
        "/api/v1/clients/{id}/add-user": {
            "patch": {
                "tags": [
                    "Clients"
                ],
                "summary": "Ajouter un compte utilisateur à un client",
                "description": "Associe un compte utilisateur à un client existant.",
                "operationId": "05cefcf3cfa744a5d6728ee7b57cf072",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte utilisateur ajouté avec succès au client",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Le client a déjà un compte utilisateur"
                    },
                    "500": {
                        "description": "Erreur lors de l'ajout du compte utilisateur"
                    }
                }
            }
        },
        "/api/v1/clients/{id}/dettes": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lister les dettes d'un client",
                "description": "Récupère la liste des dettes d'un client par son ID.",
                "operationId": "90abd66c04915e543c348482c82c4422",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des dettes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                }
            }
        },
        "/api/v1/clients/{id}/withUser": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Afficher un client avec son compte utilisateur",
                "description": "Récupère les informations d'un client et les détails de son compte utilisateur associé.",
                "operationId": "b50e4d824cca2c7fc3c53655b6e89aad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations du client récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "ClientResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "En-tête d'autorisation JWT utilisant le schéma Bearer."
            }
        }
    },
    "security": [
        {
            "BearerToken": []
        }
    ]
}